consumer/requirement.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.3
fonttools==4.49.0
gdpc==6.1.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-pubsub==2.21.2
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
idna==3.6
imageio==2.34.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.5
matplotlib==3.8.3
more-itertools==10.2.0
NBT==1.5.1
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
podman==5.0.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyGLM==2.7.1
pyglm-typing==0.2.1
pyparsing==3.1.1
python-dateutil==2.8.2
pyxdg==0.28
requests==2.31.0
rsa==4.9
scikit-image==0.22.0
scipy==1.12.0
setuptools==69.1.1
six==1.16.0
termcolor==2.4.0
tifffile==2024.2.12
urllib3==2.2.1
Werkzeug==3.0.3
wheel==0.42.0
